Fixed: Processing very long ETA from Transmission
(cherry picked from commit 9800bd6b439257e73e3545e125cd03900a3036bb)

+++ src/NzbDrone.Core/Download/Clients/Transmission/TransmissionTorrent.cs
@@ -9,7 +9,7 @@ namespace NzbDrone.Core.Download.Clients.Transmission
         public long TotalSize { get; set; }
         public long LeftUntilDone { get; set; }
         public bool IsFinished { get; set; }
-        public int Eta { get; set; }
+        public long Eta { get; set; }
         public TransmissionTorrentStatus Status { get; set; }
         public int SecondsDownloading { get; set; }
         public int SecondsSeeding { get; set; }
